Add start child selector option
By default the plugin will use the first child as the starting point for
the wizard and hide the other children. Optionally one may pass a
"startChild" selector which will filter the children of the plugin element
and use that as the root. An index variable is also created and set based
on this selector.