Add options for show/hide
Add the following options: nextShow, nextHide, prevShow, prevHide. These
are arrays that contain the parameters to pass to their respective show
hide functions. This allows users to mix and match the differnt show hide
effects. By default these are null, however if the plugin detects the
presence of the jQuery.effects module, and the user has not already set
the option, it will proivde a basic sliding effect. Hide and show are now
called via apply() so as to pass an arbitrary number of elements, this
allows the user to either use the jQuery.effects syntax, the basic time
arguments, or the easing plugin depending on their preference.